moved graph stuff to own component
